refactor ts plugin
