+++ packages/next/src/server/typescript/constant.ts
@@ -1,0 +1,17 @@
+export const NEXT_TS_ERRORS = {
+  INVALID_SERVER_API: 71001,
+  INVALID_ENTRY_EXPORT: 71002,
+  INVALID_OPTION_VALUE: 71003,
+  MISPLACED_CLIENT_ENTRY: 71004,
+  INVALID_PAGE_PROP: 71005,
+  INVALID_CONFIG_OPTION: 71006,
+}
+
+export const ALLOWED_EXPORTS = [
+  'config',
+  'generateStaticParams',
+  'metadata',
+  'generateMetadata',
+]
+
+export const LEGACY_CONFIG_EXPORT = 'config'
