+++ test/e2e/app-dir/app-alias/src/app/foo.tsx
@@ -1,0 +1,5 @@
+'use client'
+
+export default function Foo({ onClick }: { onClick: () => void }) {
+  return <button onClick={onClick}>click me</button>
+}
